+++ src/component/cpu/mc68020_moira_bridge.hpp
@@ -15,6 +15,6 @@ namespace Iris
         uint16_t read16(uint32_t addr) const override { return AddrSpace::ReadU16(addr); };
         void write8(uint32_t addr, uint8_t value) const override { return AddrSpace::WriteU8(addr, value); };  
         void write16(uint32_t addr, uint16_t value) const override { return AddrSpace::WriteU16(addr, value); }; 
-        
+    
     };
 }